+++ lib/rubocop/cop/rspec_cops.rb
@@ -5,6 +5,7 @@ require_relative 'rspec/capybara/feature_methods'
 
 require_relative 'rspec/factory_bot/attribute_defined_statically'
 require_relative 'rspec/factory_bot/create_list'
+require_relative 'rspec/factory_bot/factory_class_name'
 
 begin
   require_relative 'rspec/rails/http_status'